third_party/blink/web_tests/wpt_internal/route/routemap-003.html

<!DOCTYPE html>
<title>Set new URL, add @navigation rule</title>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-navigation-1/#at-navigation">
<script type="routemap">
  {
    "routes": [
      {
        "name": "r1",
        "pattern": [ "/*/match-later" ]
      }
    ]
  }
</script>
<style id="sheetElm">
  #target {
    width: 7px;
  }
</style>
<div id="target"></div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  promise_test(async t => {
    await new Promise(r => window.onload = () => t.step_timeout(r, 0));
    let committed = Promise.withResolvers();
    navigation.onnavigate = (event) => {
      event.intercept({
        async handler() {
          committed.resolve();
          await new Promise(resolve => requestAnimationFrame(resolve));
        }
      })
    };

    assert_equals(getComputedStyle(target).width, "7px", "Initially no match");

    // Trigger an active navigation, so that any valid "at" rule will match.
    navigation.navigate("match-later");
    await committed.promise;
    sheetElm.sheet.insertRule("@navigation (at: r1) { #target { width: 100px; } }", 1);
    assert_equals(getComputedStyle(target).width, "100px");
    sheetElm.sheet.deleteRule(1);
    assert_equals(getComputedStyle(target).width, "7px");
    sheetElm.sheet.insertRule("@navigation (at: r1) { #target { width: 200px; } }", 1);
    assert_equals(getComputedStyle(target).width, "200px");
  }, "Add and remove rules dynamically");
</script>
